requirements.txt
pandas
scikit-learn
Flask
gunicorn
psycopg2-binary